Add work experience and update project details (#21)
### TL;DR

Updated project and work experience data, added new images for work experiences.

### What changed?

- Modified the description and GitHub link for the "Habits Together" project
- Added new work experiences for McMaster Engineering Society and DeltaHacks
- Updated the description and tags for the RBC work experience
- Replaced Pekoe Development experience with McMaster University Teaching Assistant role
- Added new images for McMaster Engineering Society, DeltaHacks, and McMaster University

### Why make this change?

This update provides more accurate and up-to-date information about projects and work experiences. It showcases recent roles and responsibilities, highlighting leadership positions and academic involvement. The changes also ensure that the most relevant and current information is presented to visitors of the portfolio or resume.